refactor: create server file
